Undo for move not fixed
